+++ .gitignore
@@ -35,7 +35,7 @@ wheels/
 pip-log.txt
 pip-delete-this-directory.txt
 
-# Unit test / coverage reports
+# Unit tests / coverage reports
 htmlcov/
 .tox/
 .coverage
@@ -106,4 +106,4 @@ ENV/
 .history/
 
 # Remove file dump
-/test/dump
+/tests/dump
