+++ assets/images/3-ways-to-check-if-a-list-is-empty-in-python.jpg
