+++ tests/fonts/gadugi.ttf
